Better structure; added parabola fitting; added tetragonal and orthorhombic calculations.
Better structure of the python scripts that are used by bash scripts.
With _ at the beginning.
Display.sh will fit the parabola in $fname_output.txt and draw the
figure, also displaying the coefficients of the polynomial. Go to
c/44/0.00 to get the volume, type in the coefficient list and get
elastic constants.